chore: refactor simple orchestrator
